Move SUV2ROI code into own file
